chore(docker): introduce localstack, maildev via docker-compose
As AskGov grows, so will the sophistication of its architecture and
hence its tooling. To cater to this, introduce docker-compose to spin
up the supporting cast of services that AskGov will depend on.

This initial commit introduces localstack for S3 facilities, and
moves maildev to docker. Future changes may involve introducing more
services, eg, Elasticsearch for text search.

- Create `docker-compose.yml`, spinning up localstack and maildev
  and exposing their ports to the host
- Add `.localstack/` to .gitignore
- Replace `maildev` with `docker-compose up` in `npm run dev`, and
  remove the maildev npm package since we now use it via docker