fix(auth): return error for auth attempts with invalid emails
Previously, if there was an existing user with an invalid public service
email, we would return undefined tokens and display names from the OTP
verification endpoint. While the likelihood of such a user existing is low,
this has the added effect of making it difficult to create a shared API type
for this endpoint between the frontend and backend.

This change removes the possibility of returning undefined tokens and
display names from the OTP verification endpoint by shifting the
isOfficerEmail check earlier in the pipeline and returning an error
early for invalid emails.