feat: show link to external agency website (#421)
Add new column website to Agency, and expose to Header. Websites should contain the full path of the url, including protocol (http), otherwise the link will still stay in askgov. The button will display only the hostname instead of the full url to maintain legitimacy (.gov) and brevity.

If the website is empty or at homepage, the link does not show up.